+++ src/bin/src/main.rs
@@ -59,10 +59,7 @@ async fn main() -> Result<(), Box<dyn Error>> {
     let (tx_events, rx_events) = flume::unbounded();
     let (tx_events_router, rx_events_router) = flume::unbounded();
 
-    // TODO after everything has been migrated, keep only the AppConfig and get rid of AppState
-    //  and all constants
-    info!("Initializing AppConfig");
-
+    info!("Initializing Config");
     let (rauthy_config, node_config) = RauthyConfig::build(
         config_file,
         tx_email.clone(),