+++ src/models/src/entity/refresh_tokens.rs
@@ -57,8 +57,8 @@ impl RefreshToken {
         exp: i64,
         scope: Option<String>,
         // TODO should we even save mfa for refresh tokens?
-        // even if the original token has been issued with mfa, the refresh
-        // token not really is, because it can be given without user interaction.
+        //  even if the original token has been issued with mfa, the refresh
+        //  token not really is, because it can be given without user interaction.
         is_mfa: bool,
         session_id: Option<String>,
     ) -> Result<Self, ErrorResponse> {