+++ src/models/src/events/event.rs
@@ -149,7 +149,7 @@ pub enum EventType {
     NewUserRegistered,
     NewRauthyAdmin,
     NewRauthyVersion,
-    PossibleBruteForce, // TODO
+    PossibleBruteForce, // currently unused - kind of covered by IpBlacklisted
     RauthyStarted,
     RauthyHealthy,
     RauthyUnhealthy,
@@ -159,6 +159,9 @@ pub enum EventType {
     Test,
     BackchannelLogoutFailed,
     ScimTaskFailed,
+    ForcedLogout,
+    UserLoginRevoke,
+    SuspiciousApiScan,
 }
 
 impl Default for EventType {
@@ -176,7 +179,7 @@ impl Display for EventType {
             Self::IpBlacklistRemoved => write!(f, "IP blacklist removed"),
             Self::JwksRotated => write!(f, "JWKS has been rotated"),
             Self::NewUserRegistered => write!(f, "New user registered"),
-            Self::NewRauthyAdmin => write!(f, "New rauthy_admin member"),
+            Self::NewRauthyAdmin => write!(f, "New Rauthy_admin member"),
             Self::NewRauthyVersion => write!(f, "New Rauthy App Version available"),
             Self::PossibleBruteForce => write!(f, "Possible brute force"),
             Self::RauthyStarted => write!(f, "Rauthy has been restarted"),
@@ -188,6 +191,9 @@ impl Display for EventType {
             Self::Test => write!(f, "TEST"),
             Self::BackchannelLogoutFailed => write!(f, "Backchannel logout failed"),
             Self::ScimTaskFailed => write!(f, "SCIM task failed"),
+            Self::ForcedLogout => write!(f, "Forced user logout"),
+            Self::UserLoginRevoke => write!(f, "User revoked illegal login"),
+            Self::SuspiciousApiScan => write!(f, "Suspicous API scan"),
         }
     }
 }
@@ -214,6 +220,9 @@ impl From<rauthy_api_types::events::EventType> for EventType {
                 Self::BackchannelLogoutFailed
             }
             rauthy_api_types::events::EventType::ScimTaskFailed => Self::ScimTaskFailed,
+            rauthy_api_types::events::EventType::ForcedLogout => Self::ForcedLogout,
+            rauthy_api_types::events::EventType::UserLoginRevoke => Self::UserLoginRevoke,
+            rauthy_api_types::events::EventType::SuspiciousApiScan => Self::SuspiciousApiScan,
         }
     }
 }
@@ -238,6 +247,9 @@ impl From<EventType> for rauthy_api_types::events::EventType {
             EventType::Test => Self::Test,
             EventType::BackchannelLogoutFailed => Self::BackchannelLogoutFailed,
             EventType::ScimTaskFailed => Self::ScimTaskFailed,
+            EventType::ForcedLogout => Self::ForcedLogout,
+            EventType::UserLoginRevoke => Self::UserLoginRevoke,
+            EventType::SuspiciousApiScan => Self::SuspiciousApiScan,
         }
     }
 }
@@ -262,6 +274,9 @@ impl EventType {
             Self::Test => "TEST",
             Self::BackchannelLogoutFailed => "BackchannelLogoutFailed",
             Self::ScimTaskFailed => "ScimTaskFailed",
+            Self::ForcedLogout => "ForcedLogout",
+            Self::UserLoginRevoke => "UserLoginRevoke",
+            Self::SuspiciousApiScan => "SuspiciousApiScan",
         }
     }
 
@@ -287,6 +302,9 @@ impl EventType {
             EventType::Test => 14,
             EventType::BackchannelLogoutFailed => 15,
             EventType::ScimTaskFailed => 16,
+            EventType::ForcedLogout => 17,
+            EventType::UserLoginRevoke => 18,
+            EventType::SuspiciousApiScan => 19,
         }
     }
 }
@@ -311,8 +329,14 @@ impl From<String> for EventType {
             "TEST" => Self::Test,
             "BackchannelLogoutFailed" => Self::BackchannelLogoutFailed,
             "ScimTaskFailed" => Self::ScimTaskFailed,
+            "ForcedLogout" => Self::ForcedLogout,
+            "UserLoginRevoke" => Self::UserLoginRevoke,
+            "SuspiciousApiScan" => Self::SuspiciousApiScan,
             // just return test to never panic
-            _ => Self::Test,
+            s => {
+                error!("EventType::from() for invalid String: {s}");
+                Self::Test
+            }
         }
     }
 }
@@ -343,6 +367,9 @@ impl From<i64> for EventType {
             14 => EventType::Test,
             15 => EventType::BackchannelLogoutFailed,
             16 => EventType::ScimTaskFailed,
+            17 => EventType::ForcedLogout,
+            18 => EventType::UserLoginRevoke,
+            19 => EventType::SuspiciousApiScan,
             _ => EventType::Test,
         }
     }
@@ -453,6 +480,12 @@ impl From<&Event> for Notification {
                 value.data.unwrap_or_default(),
                 value.text.as_deref().unwrap_or_default()
             )),
+            EventType::ForcedLogout => Some(format!(
+                "User `{}` was force-logged-out",
+                value.text.as_deref().unwrap_or_default()
+            )),
+            EventType::UserLoginRevoke => value.text.clone(),
+            EventType::SuspiciousApiScan => value.text.clone(),
         };
 
         Self {
@@ -631,7 +664,11 @@ impl Event {
 
     pub fn backchannel_logout_failed(client_id: &str, user_id: &str, retries: i64) -> Self {
         Self::new(
-            EventLevel::Critical,
+            RauthyConfig::get()
+                .vars
+                .events
+                .level_backchannel_logout_failed
+                .clone(),
             EventType::BackchannelLogoutFailed,
             None,
             Some(retries),
@@ -646,6 +683,16 @@ impl Event {
             Some(ip),
             None,
             None,
+        )
+    }
+
+    pub fn force_logout(user_email: String) -> Self {
+        Self::new(
+            RauthyConfig::get().vars.events.level_force_logout.clone(),
+            EventType::ForcedLogout,
+            None,
+            None,
+            Some(user_email),
         )
     }
 
@@ -709,6 +756,27 @@ impl Event {
         )
     }
 
+    pub fn user_login_revoke(
+        user_email: &str,
+        bad_ip: IpAddr,
+        bad_location: Option<String>,
+    ) -> Self {
+        let loc = bad_location.as_deref().unwrap_or("Unknown Location");
+        let text = format!("User `{user_email}` revoked illegal login from {bad_ip} ({loc})");
+
+        Self::new(
+            RauthyConfig::get()
+                .vars
+                .events
+                .level_user_login_revoke
+                .clone(),
+            EventType::UserLoginRevoke,
+            Some(bad_ip.to_string()),
+            None,
+            Some(text),
+        )
+    }
+
     pub fn rauthy_started() -> Self {
         let text = format!("Rauthy has been started on host {}", get_local_hostname());
         Self::new(
@@ -765,7 +833,11 @@ impl Event {
 
     pub fn scim_task_failed(client_id: &str, action: &ScimAction, retries: i64) -> Self {
         Self::new(
-            EventLevel::Critical,
+            RauthyConfig::get()
+                .vars
+                .events
+                .level_scim_task_failed
+                .clone(),
             EventType::BackchannelLogoutFailed,
             None,
             Some(retries),
@@ -784,6 +856,23 @@ impl Event {
             Some(ip.to_string()),
             None,
             None,
+        )
+    }
+
+    pub fn suspicious_request(path: &str, ip: IpAddr, location: Option<String>) -> Self {
+        let loc = location.as_deref().unwrap_or("Unknown Location");
+        let text = format!("Suspicious request to '{path}' from {ip} ({loc})");
+
+        Self::new(
+            RauthyConfig::get()
+                .vars
+                .events
+                .level_suspicious_request
+                .clone(),
+            EventType::SuspiciousApiScan,
+            Some(ip.to_string()),
+            None,
+            Some(text),
         )
     }
 
@@ -847,7 +936,7 @@ impl Event {
                     self.text.as_deref().unwrap_or_default()
                 )
             }
-            // PossibleBruteForce is not yet implemented / recognized
+            // PossibleBruteForce is currently not in use
             EventType::PossibleBruteForce => String::default(),
             EventType::RauthyStarted => self.text.clone().unwrap(),
             EventType::RauthyHealthy => self.text.clone().unwrap(),
@@ -879,6 +968,14 @@ impl Event {
                     self.text.as_deref().unwrap_or_default()
                 )
             }
+            EventType::ForcedLogout => {
+                format!(
+                    "User `{}` was force-logged-out",
+                    self.text.as_deref().unwrap_or_default()
+                )
+            }
+            EventType::UserLoginRevoke => self.text.clone().unwrap_or_default(),
+            EventType::SuspiciousApiScan => self.text.clone().unwrap_or_default(),
         }
     }